Started work on the shop, the shop won't be a fixed-type of shop, it'll be dynamic meaning you'll be able to customize it however you'll want, with as many categories and pages, as well as store items.
Created a dynamic quantity page that will allow developers to use it, this will be featured in the shop menu as well.

Fixed many bugs that did not allow you to open some menus.

Made the gens more performant, so the plugin should be more smoother, I would like to remind people that this plugin isn't ready for production use and is mearely to be used to experiment until the beta release, I appreciate all feed-back given to me about the plugin, if you find a bug, please message me on Discord Frostingly™#6666!